Module masquerade_proxy::common

Structs

ToSend

QUIC packet

Enums

ConfigError

Content

HTTP/3 message types.

Constants

MAX_DATAGRAM_SIZE

MAX_INT_LEN_1 🔒

MAX_INT_LEN_2 🔒

MAX_INT_LEN_4 🔒

MAX_VAR_INT

Functions

decode_var_int

decode_var_int_get_length

encode_var_int

get_next_ipv4

Gets the next IPv4 address If the next address isn’t allowed by the netmask an error is returned.

hdrs_to_strings

hex_dump

interrupted

make_qlog_writer

Makes a buffered writer for a qlog.

mint_token

Generate a stateless retry token.

send_h3_dgram

split_ip_prefix

Parses an IP and splits off the prefix If the prefix was not found or if it wasn’t possible to parse the second part of the 2 tuple will be None.

validate_token

Validates a stateless retry token.

would_block