Function masquerade_proxy::common::split_ip_prefix

pub fn split_ip_prefix(ip: String) -> (String, Option<u8>)

Parses an IP and splits off the prefix If the prefix was not found or if it wasn’t possible to parse the second part of the 2 tuple will be None.

Examples

    assert_eq!(split_ip_prefix("192.168.0.0/24".to_string()),
                ("192.168.0.0".to_string(), Some(24)));
                
    assert_eq!(split_ip_prefix("192.168.0.0".to_string()),
                ("192.168.0.0".to_string(), None));
     
    assert_eq!(split_ip_prefix("192.168.0.0/ab".to_string()),
                ("192.168.0.0".to_string(), None));