Function masquerade_proxy::common::validate_token

pub fn validate_token<'a>(
    src: &SocketAddr,
    token: &'a [u8],
) -> Option<ConnectionId<'a>>

Validates a stateless retry token.

This checks that the ticket includes the "quiche" static string, and that the client IP address matches the address stored in the ticket.

Note that this function is only an example and doesn’t do any cryptographic authenticate of the token. It should not be used in production system.