Module masquerade_proxy::connect_ip::util

Structs

HandleIPError

IPError

QUICStreamError

UdpBindError

Enums

Ipv4CheckError

Constants

TCP_ID 🔒

UDP_ID 🔒

Functions

calculate_tcp_udp_checksum

Recalculate the checksum of a udp or tcp packet References:

check_ipv4_packet

checksum_add 🔒

encapsulate_ipv4

Creates a ToSend struct containing an IP packet

get_ip_header_length

Return the header length of a given packet slice The header len is at the second nibble of a ipv4 packet Header length is given in 32bit words. A header value of 0b1111 = 15, 15*32=480bit=60byte

get_ip_version

Returns the version of the ip packet slice, given in the first nibble

get_ipv4_hdr_checksum

Read the checksum of a ipv4 packet Does not calculate the checksum, only reads it from the header!

get_ipv4_pkt_dest

Reads the destination addr of a given IPv4 packet. Warning: This does NOT check if this is a valid IP packet, or even if the pkt is long enough.

get_ipv4_pkt_source

Reads the source addr of a given IPv4 packet. Warning: This does NOT check if this is a valid IP packet, or even if the pkt is long enough.

get_ipv4_ttl

Return the ttl of a given packet slice

recalculate_checksum

Recalculates the checksum of a ipv4 packet. If the payload is TCP or UDP we also recalculate that checksum.

set_ipv4_pkt_destination

Sets the destination ip address of a given IPv4 buffer to the given adress. Warning: This does NOT check if this is a valid IP packet, or even if the pkt is long enough.

set_ipv4_pkt_source

Sets the source ip address of a given IPv4 buffer to the given adress. Warning: This does NOT check if this is a valid IP packet, or even if the pkt is long enough.

update_ipv4_checksum

Updates the checksum of a IPv4 header to be correct.