Module masquerade_proxy::connect_ip::client

Structs

ClientConfig

Used for reading and applying user defined configurations. Should be read in at start of client.

ConnectIPClient

ConnectIpInfo

Infos about the CONNECT-IP session Includes converters for local ip’s to destination ip’s (and the other way around)

IpMessage 🔒

Holds an ip packet and its direction

QuicStream

Useful for holding some basic data for a HTTP3 Stream

Enums

Direction

Information about packets, wether they are to be sent to the server or to the client. Generally, packets that were received from the server (from the QUIC connection) are labeled as ToClient, and packets from the TUN device as ToServer

Functions

generate_cid_and_reset_token

Generate a new pair of Source Connection ID and reset token.

handle_ip_connect_stream 🔒

Initiates the CONNECT-IP request. Arguments: * http3_sender: To send the http3 requests to * response_receiver: Receives responses on sent http3 requests from QUIC handler * config: Client configuration determined by user

ip_dispatcher_t 🔒

Receives ready-to-send ip packets and then sends them to the TUN device. Arguments: * ip_dispatch_reader: Waits for messages from ip handler thread * writer: Writes messages to the TUN device

ip_handler_t 🔒

Receives IP Packets from rx. Will then handle these packets accordingly and send messages to quic_dispatcher_t Arguments: * ip_recv: Received IPMessage objects from the QUIC side or the TUN receiver * conn_info_recv: Channel used for sending the connection config from QUIC side * http3_dispatch: For sending encapsuled IP packets to the QUIC side * ip_dispatch: For sending IP messages from the QUIC side to the TUN device * device_addr: Address injected into packets directed to the system

ip_receiver_t 🔒

Receives raw IP messages from a TUN.the Will send received messages to the ip_handler_t Arguments:

quic_conn_handler 🔒

A general handler of a quic connection. Will set up the connection and the connect_ip HTTP/3 connection. Then receives ip address and route advertisement. Sends these infos to quic_dispatcher and ip_handler After that goes into loop, waits for new messages and handles them accordingly. Sends resulting messages to either ip_handler_t or quic_dispatch_t.

set_client_ip_and_route 🔒

Basic commands for setting up the TUN interface Will also route the traffic according to the allowedIp’s parameter.